move API service to shared + make headers optional
