+++ app/mobile-content/shared/mobile-content-api.service.ts
@@ -1,14 +1,14 @@
 import { Injectable } from '@angular/core';
 import { AuthHttp } from 'angular2-jwt';
 
-import { ConfigService } from '../config.service';
+import { ConfigService } from '../../config.service';
 
 /**
  * Service for sending request to the mobile content API.
  */
 @Injectable()
 export class MobileContentApiService {
-  mobileContentApiUrl: string;
+  private mobileContentApiUrl: string;
 
   constructor(private http: AuthHttp,
               private config: ConfigService) {}
@@ -26,7 +26,7 @@ export class MobileContentApiService {
    * @param headers additional headers
    * @returns {Observable<Response>}
    */
-  public getUrl(apiUrl: string, headers: any) {
+  public getUrl(apiUrl: string, headers: object = {}) {
     this.setUrl();
     return this.http.get(this.mobileContentApiUrl + apiUrl, headers);
   }
@@ -39,7 +39,7 @@ export class MobileContentApiService {
    * @param headers additional headers
    * @returns {Observable<Response>}
    */
-  public postUrl(apiUrl: string, data: string, headers: any) {
+  public postUrl(apiUrl: string, data: string, headers: object = {}) {
     this.setUrl();
     return this.http.post(this.mobileContentApiUrl + apiUrl, data, headers);
   }
@@ -52,7 +52,7 @@ export class MobileContentApiService {
    * @param headers additional headers
    * @returns {Observable<Response>}
    */
-  public putUrl(apiUrl: string, data: string, headers: any) {
+  public putUrl(apiUrl: string, data: string, headers: object = {}) {
     this.setUrl();
     return this.http.put(this.mobileContentApiUrl + apiUrl, data, headers);
   }
@@ -64,7 +64,7 @@ export class MobileContentApiService {
    * @param headers additional headers
    * @returns {Observable<Response>}
    */
-  public deleteUrl(apiUrl: string, headers: any) {
+  public deleteUrl(apiUrl: string, headers: object = {}) {
     this.setUrl();
     return this.http.delete(this.mobileContentApiUrl + apiUrl, headers);
   }
